CityBuild-Server-Manager/requirements.txt
mcrcon==0.6.1
colorama==0.4.6
python-dotenv==1.0.0
